api/requirements.txt
Flask
PyJWT
Flask-SQLAlchemy
psycopg2
